feat: remove dev tools and tools controllers
- Remove dev tools and tools controllers
- Replace the render using an AngularJS template by a ReactJS component
- Create ToolsRouter component to manage the routing of Dev Tools and
  Ruleset Test views
- Move logic and render of dev tools to a ReactJS component
- Remove unused definitions of controllers and variables in AngularJS
